meson: Use Meson to build CLI and libhb
Extends the existing Meson build system for LinGUI to build the
libhb and HandBrakeCLI components as well. The new make targets
are prefixed with hb, for example hb.build and hb.xclean.
Although there is now a meson.build file in the root directory,
you still need to run configure with the existing build flags to
set up the project.